ログの確認と実装済み処理のリファクタリングをした #45
