+++ DietApp/Model/Enum/Direction.swift
@@ -1,0 +1,11 @@
+//
+//  Direction.swift
+//  DietApp
+//
+//  Created by 川島真之 on 2023/07/08.
+//
+
+enum Direction {
+  case next
+  case previous
+}
