+++ DietApp/TopPage/NavigationController+Orientation.swift
@@ -5,7 +5,7 @@
 //  Created by 川島真之 on 2023/06/20.
 //
 
-import Foundation
+
 import UIKit
 
 extension UINavigationController {
